+++ RELEASE_NOTES.md
@@ -1,12 +1,10 @@
 Release Notes
 =============
 
-## 1.7.0
-* CosmosDb: Add support for serverless capacity mode.
-
 ## 1.6.23
 * ContainerApps: Adds support for [containerApps](https://docs.microsoft.com/azure/container-apps/overview).
 * WebApps/Functions: Added support for .NET 6 runtimes with new Runtime.DotNet60.
+* CosmosDb: Add support for serverless capacity mode.
 
 ## 1.6.22
 * Log Analytics: Add CustomerId configuration member to Log Analytics